Mitochondrial Transcripts Filter

Automatically identify an appropriate threshold for subsetting a Seurat object by mitochondrial percentage (fraction of total transcripts beginning with "MT-". MitoFilter first identifies the median mitochondrial percentage value five standard deviations above the median for a given sample. MitoFilter then subsets a Seurat object, retaining only cells which have a lesser mitochondrial percentage than the cutoff (median +5 SD). The object is returned containing only cells below threshold, as well as only cells with greter than 100 unique RNA transcripts.

MitoFilter(obj = NULL)

Arguments

obj
The vector of variable features used to construct the PCs

mtx.use
The Seurat object to subset. This object must already contain a metadata feature "percent.mito" containing the total mitochondrial percentage, by cell. This metadata feature can be added by the "PercentageFeatureSet" function.

Value

Returns a filtered Seurat object with only cells passing the mitochondrial percentage filter, as well as containing greater than 100 unique RNA transcripts.

Examples

my_seurat_object <- MitoFilter(obj = my_seurat_object)
#> Error in is.data.frame(x) object 'my_seurat_object' not found
my_seurat_object <- my_seurat_object %>%
                    MitoFilter()
#> Error in my_seurat_object %>% MitoFilter() could not find function "%>%"